Major overhaul of DOs I/O and AppDataObjects classes
From what I could read in the current code it wasn't clear to me how writing into a DataObject occurred, how reading happened, how these activities were tied to the state machine of DataObjects, which carried a number of consequences. In particular, I found confusing to understand how AppDataObjects fit into the current design, because it wasn't clear to me where they were reading data from and where were they writing data to. After having a second reading to the design documents and artifacts I thought I would give it a try and re-code the existing classes to conform better with the current design.

First of all more status enforcement has been added. For instance, writing to a DO cannot occur after it's transitioned to COMPLETED, and opening or reading cannot occure before the DO is COMPLETED. For reading, one must open() the DO, which returns a descriptor which is local to that opening action (and thus it allows for multiple, concurrent openings). The descriptor must thus be used while reading the DO, which must be closed at the end of the operation. An easier interface for the 'with' keyword support would probably be nice. Also, to allow for oneliner full-reading operations a new doutils module has been added, which contains a method that returns the full contents of a DO. Some DO classes existed already, and some have been added. Thus we now have the InMemoryDataObject, FileDataObject, NgasDataObject and ContainerDataObject classes, which basically implement different storage mechanism.

On a separate side we have Consumers. Consumers are added to DOs via the addConsumer() call, and are called back when the DO's state changes to COMPLETED, thus effectively being notified when the DO data is fully available. One particular type of Consumer is the AppConsumer, a class that runs a command or executes some code using the dataObject that changed to COMPLETE, writing the result of the computation into itself using a write() method, and marking itself as COMPLETED. When a class subclasses both the AbstractDataObject and the AppConsumer it then becomes a DataObject type that reacts upon completion of another DO, and that represents the result of the computation that it's supposed to run. I've modified the previous ChecksumAppDataObject to be then a simpler CRCResultConsumer, and then provided three subclasses used to store this CRC result in different ways: FileCRCResultDataObject, NgasCRCResultDataObject and InMemoryCRCResultDataObject. This is a simple example of an AppDataObject, but there are a few more in the test_data_object test module. All of them still represent a single DataObject though; the next step is to mix the AppConsumer with a ContainerDataObject, which would represent an application that produces many outputs.

Regarding tests, those under test_data_object and test_dlm all expect for one that still needs some attention, but is almost there.

Finally, some small code beautification has been done, mostly changing prints by loggers and removing unused imports.