Added the SocketListener class
With this class, and the new InMemorySocketListenerDataObject and FileSocketListenerDataObject mixes we have DOs that listen for incoming data in a given (TCP) socket and write it into themselves, moving themselves finally to the COMPLETED state when the client closes the connection. A small test that exercises this logic has been added.

In the test we are reusing the EvtConsumer class