Luigi integration mainlined
Now that integration with luigi has been working for some time, both for monitoring and executing DOs, I've moved the relevant luigi Tasks to the main dfms package instead of having them residing in the test directory.

Together with this movement I took all the different graphs used for luigi tests and put them in a graphsRepository module, containing functions that generate each of the graphs. This makes the graph definitions reusable, and I'm actually currently using them as the input for unit testing that the luigi integration Tasks actually work in test_luigi.py.

Finally, some more auxiliary code has been moved to doutils to navigate the DO graph and perform operations on its nodes.