DOM supports daemon mode
Now DOM instances (which can be launched via the dfmsDOM command-line script) support their own daemonization, which will write the PID of the process to a file in the user's home directory. More than one DOM instance can be created in the same host by specifying different ports used by the DOM to listen for requests.