DOM creating graphs, inter-DOM graphs working
The DOM now has a method to accept a graph specification as an input, which will create the full graph contained in the specification (i.e., all its DOs), and make them available via the session's Pyro Daemon. All the DOs of the graph are locally connected. One can then feed different graphs to different DOMs, have them create their graphs, and then get a Proxy reference to the end/root nodes of the graph and establish inter-DOM relationshipts between DOs created in different DOMs. For this to work the addChild method of the ContainerDataObject had to be changed to make it pyro-friendly (but still without any reference to Pyro itself). Automatic tests check that the producer/consumer and the parent/child relationships are working. Since producer/consumer is working, immediateProducer/immediateConsumer is assumed to be also working.