Updating the code base after discussions on the architecture of the DFMS
The core concept of the Application DataObject has been discussed thoroughly during the last days and we have agreed the way it should be implemented. During the same discussions we made also clear what the Container DataObject means and how to implement barriers.

This commit reflects all the changes we discussed during these conversations. Although all the code base has been updated, and all the tests are happily passing, there are still a number of questions that need to be answered now that we have switched to the new model. Personally the main ones are:
 * Should ContainerDataObjects implement the I/O methods? It seems it will be necessary, and it will also mean that Container DataObjects don't only point to their children, but also need to store some bits themselves (although a new children could be used for that).
 * The lifecycle of the AppDataObjects is still unclear.