Small fixes and improvements in dfms.dom package
A new command-line flag allows the user to disable the DLM for a given DOM instance.
Now also the graph shows both the DO type (app, container, etc) in the color of the border of the node representing it, as well as the status, with the color of the left portion of the inner area of the node.