'Normal' consumers execute on a different thread
Now 'normal' consumers (i.e., those reacting to the change of a DO to its COMPLETED state) are executed asynchronously in a different execution thread. Although the change is simple, it meant that several tests involving chained reactions had to be adjusted to make sure they take this multi-threaded situation into account.