Improved explicit session support
Previously the DataObjectManager handled independent sessions by itself, maintaining several dictionaries and lists with objects belonging to different sessions in them. It also took care of creating DataObjects, starting the Pyro daemons, etc. I thus introduced a new Session class, which takes care of all aspects related to a session, making the DOM simpler. The DOM still bind DOs created by the Sessions with its DLM, so no functionality has been lost.

A session now obeys a simple state machine, with an initial state where the graph to be run can be constructed piece by piece, by add graph specifications to the session, and then run, which moves the session to the deploying and running state. The session also uses the luigi tasks now to run the graph, and after the execution finishes the session is moved to finished. This granularity is also exposed by new methods in the DOM, and in the REST API that lies on top of it.