Terminology change: 'immediate' --> 'streaming'
The previously called 'immediateConsumers' and 'immediateInputs' have been renamed to 'streamingConsumers' and 'streamingInputs' for sake of clarity, since that's the term we use in our design and discussions normally.

Together with this change I also removed the 'immediateProducer' concept from the AbstractDataObject, since in the rest of the code we assume applications don't differentiate between 'streaming' or 'normal' outputs actually.