Adding graph templates
Graph templates are kept in a repository (currently a hardcoded module with a given number of functions, but probably will need to be a per-user directory with python modules living inside). They contain functions that yield graph specifications, and which accept (optionally optional) parameters. These templates are listed by the DOM, and can also be used to construct a physical graph in an existing session. This new functionality is also exposed through the current REST interface.