Changing how apps trigger DO completeness.
So far each App was manually calling setComplete() on its outputs once it had finished writing into them. This implies that the application has total control over the output, and that it's its only writer. This is not the case if we want to support multiple apps writing to a single output. To accomodate to this scenario, now the App signals when it has finished its execution (either with an error or not), and the output reacts to this signal moving itself to the COMPLETED state. Using this paradign multiple applications writing to a single output can signal the end of their executions individually, and the output DO will decide when to move to the COMPLETED state after all its producers have finished (not yet implemented).