New ngaslite module
The ngaslite module is used by the dfms whenever the NGAS libs are not found in the system, but NGAS access is still necesary. An NgasLiteIO class exposes its functionality in the form of an IO class, which means that it can be used seamlessly by users of the IO classes, like DataObjects and applications like the NgasArchivingApp.

The NgasArchivingApp uses the NgasIO class by default to store its inputDO in NGAS, but if not found it will use the NgasLiteIO class instead. This fixes the test failure seen in Jenkins after the last commit.