Base class for external archiving applications
Separating the NgasArchivingApp into a base class and a ngas-specific class allows for other types of storages to be plugged in. These two classes finally make up a small example on how an application put at the end of the physical graph (or more precisely, right after the necessary DataObjects) can automatize the data migration and archiving processes.

This code is currently being tested by the test.test_dom.TestREST.test_fullRound method, where a DOM is started, a physical graph is created in a new session, two archiving applications are created and connected to the graph, and the graph is executed. The test finally checks that the contents of the objects in the remote NGAS server matches the expected output, which is unique to each test run.