DockerApp plus first basic test
DockerApp wraps the execution of a docker-ized application. It pulls the necessary image at DO deployment time, and executes the command in the docker container at runtime. It makes sure that the inputs and outputs are of the correct type (otherwise they currently cannot be used with the DockerApp), and ensures that the owner of the outputs is the same user running the dfms framework (and not the root user in the container). The test checks all these aspects of the application. The only missing bit is to actually check for the presence of the docker daemon.