Use the correct user inside Docker containers
We already don't create an 'r' user inside the container because there already exists a user with the needed UID. On the other hand, we were always running things as the 'r' user with the 'su' command, which would be wrong if an already-existing user with the desired UID is not called 'r'. We thus always calculate the username that corresponds to the UID we need to use.