Luigi calls directly .execute()
When a DO's executionMode was previously set to EXTERNAL, the luigi task for that DO would then "execute" it by calling the callback methods for each of the DO's input. This was a bit of a hack, and now that we have a separate .execute() method we can call it directly, since it takes care of the execution and status tracking already.

During this exercise I also realized that luigi can only execute via .execute() (or via the previous method) not all AppDataObjects, but only BarrierAppDataObjects, so I changed that. I also moved the (temporary) change to COMPLETED inside the .execute() method in order to have it centralized and consistent.