Adding a rdbms-based DLC registry interface.
In the past an InMemoryRegistry existed as part of the current DLC prototype, which obeyed a Registry interface (addDataObjectInstance, recordNewAccess, etc.). A separate implementation of the same interface has been included now in the prototype, which obeys the same Registry interface (and therefore they are fully interchangeable), and that uses any DB-API v2.0 compliant python module to perform SQL operations. A unit test has been also included, using the sqlite3 python module and a file-based database.

This DB implementation was mentioned in PRO-109, and probably gives some initial information as to what is needed for PT-254