Reorganized DOLinkType->stuff maps
DOLinkType is the main way to identify a relationship type. We have now two separate dictionaries based on these values: one to express the DataObject method/property related to that DOLinkType that lives in the data_object module (because in that module we already have all that information, and because we can then reference it from other modules that sit above), and one expressing the doSpec keys for each link type, which lives in the graph_loader module. This way we separate concerns and increase reusability.