Initial version of the DataIslandManager
The DIM basically manages a number of DOMs, which are assumed to be one per node. In other words, in manages a number of hosts. It brings up DOMs via ssh if they are not up yet, and communicates with them by getting a Pyro proxy via the Pyro NameServer. Its methods shadow those of the DOM of course.

One of the key aspects of the DIM is that it receives the physical graph for an entire island, and therefore it has some extra requirements, like the 'location' field being set in all doSpecs.

Some basic unit tests check that the DIM is really performing actions on the DOMs it's supposed to manage.