Graph BFD with optional filtering.
The function given as an argument to the dfms.doutils.breadFirstTraverse (sic) method can now accept either one or two arguments. If two arguments are accepted, then the second argument is the list of nodes connected to the current visited node that will be visited later. This list can be modified within the function where the node is visited, effectively filtering out parts of the graph if necessary. Both normal functions and class instance methods are accepted