More methods and tests for the DIM
getGraphStatus, getGraph and getSessionStatus have been added to the DIM, which in turn gathers the results of the individual DOMs. The gathering happens in separate threads. Unit tests have been added for the new methods.

Additionally, proper error propagation has been added at the DIM level when parallel executions take place to invoke DOM methods. Exceptions are collected per-DOM, and if any is found after all threads finish a new exception is raised in the DIM method caller thread to show that there was an error in some of the per-DOM threads.