Simplifying the Session.addGraphParts method
Before we allowed a number of types for the argument of Session.addGraphParts, which in the end had to be a list of dictionaries. We now allow only one type, simplifying the code of the Session class, and at the same time removing unnecessary functions from the graph_loader module.