Keep looking for DOM in the NS
Otherwise, it's possible that the DOM has started, but haven't yet completed its registration with the NS, in which case the DIM won't be able to find it.

It's strange that this happens anyway, because both the DIM and the DOM register with the NS before starting the pyro daemons that host them, and on the other hand the DIM checks that it can connect to the daemon when starting up the DOMs, raising an exception otherwise. The fact that the DIM doesn't get an exception after starting the DOM means that the DOM should be registered with the NS already.