Allow to specify DOs that start in COMPLETED state
When a graph is deployed the user can give a list of UID to specify which DOs should be transitioned to COMPLETED at the moment of deploying the graph. When the list is given to the DOM the DOs are moved to COMPLETED right after they are created. When the deployment is done trhough the DIM, it doesn't pass down the list to the individual DOMs, but instead deploys all the graphs, creates all the inter-DOM DO connections, and only then it (via Pyro proxies) performs the transitions to COMPLETED.