SocketListener is now SocketListenerApp
This change means that now there is no longer a special 'socket' DataObject, which was still a mixture of an application DO *and* of a simple storage DO. Instead, the new SocketListenerApp simply reads the data written into the socket, and writes it into its output DOs, making it simply one more app. The new SocketListenerApp also opens its socket at run() time, which means that the DO nows needs to be somehow triggered after being created to have the socket open. Since the DOM and the DIM already had the 'completed' feature during the deployment of a graph, this mechanism is also used to start BarrierApps in general. SocketListenerApp derives from BarrierApp (although it accepts no input, as it is designed to receive its input data via a socket).

This new SocketListenerApp lives in the dfms.apps.socket_listener module. Likewise, I moved the CRCAppDataObject from the dfms.data_object module into the new dfms.apps.crc module. This way now the dfms.data_object contains no specific applications.

These two changes meant mostly that all the graphs in the examples and tests needed to be revised to split the previous InMemorySocketListenerDataObjects into two DOs: one SocketListenerApp and an InMemoryDataObject as its output.