Better user and binding handling of DockerApps.
DockerApps now accept two options that have to do with the user that will run the container process. One specifies the user used to run the container, while the other specifies a user to be created if necessary in the running container (assuming the container is run as root), which then root will switch to for running the command.

On the volume binding side we map the dirname of the outputs instead of their full path. This way they are not created by docker, avoiding also the custom hack needed for file outputs.

All unit tests still work, but more importantly the chiles example also works now.