Corrected and improved DOM session display.
Sessions in the DOM web interface now show the graph status correctly, whereas before it was incorrectly matching classes with nodes in the graph (although there is a caveat: if different graph parts are appended and the interface is aware of them at different graph query times it will still break). Now it also shows the currently running application, making it more obvious where the current execution is (it previously also showed the currently writing DO, but that state is almost never reached in the real examples).