Removed need to use Pyro's NameService.
Since the DIM knows the port where the DOMs should be running, and assigns an ID to them, we can use that ID as the objectId with which the DOM register themselves into their Pyro Daemons. That way the DIM can also use those IDs to look for the DOMs, making it unnecessary to use the NameService.

This heavily implies that there is at most one DOM and one DIM per host. Their ports are also still fixed.