Don't do consumer.dataObjectCompleted on a thread
Calls to consumer.dataObjectCompleted are not always expensive. In some cases the call simply increases a counter on the consumer side, like in the case of the BarrierAppDataObjects. Instead, the thread is created in the consumer side, who has more knowledge about its behaviour and the need of actually creating a thread or not.