New ScpApp
The new ScpApp copies over the contents from one DataObject to another, which can reside in different nodes. The information of where a DataObject resides is now stored in their `node` property, which defaults to 127.0.0.1 but is a mandatory field on the physical graph when submitted to the DIM.

The application itself hasn't been tested yet, but its underlying code (under dfms.remote) has been manually, and successfully, tested.