Rethinking events
All the inter-DROP communication is not done *without* events. This is to make more clear the different interactions between DROPs, without the complexity of having producer-side event handlers that would do some filtering before calling a receiving-side DROP method (e.g., how it was done in the AbstractDROP.addConsumer method). The current remaining users of events are the DLM, a small piece of code within the luigi integration, and a few lines in the DockerApp.

I also simplified the EventBroadcaster classes, collapsing them to a single one, called EventFirer now. The EventFirer class now serves as the parent class for the AbstractDROP, so it saves us some lines of code as well.

Finally, subscription is object-based, not callback-based. Objects need to have a handleEvent method that receives the event. This will make it easier in the future to subscribe using remote object proxies if needed.