Support for multi-level management hierarchy.
Until now we had a fixed two-level management hierarchy where the DataIslandManager managed DROPManagers, which in turn held the DROPs directly. By turning the DataIslandManager into a more generic CompositeManager (without many changes actually, mostly variable/method renaming) we have a true multi-level hierarchy support, where the CompositeManager manages other managers, regardless if they are CompositeManagers of DROPManagers.

From this new CompositeManager we now derive the good, old DataIslandManager, and also a new MasterManager, which manages DataIslandManagers. This new MasterManager is a new concept that appears in the SDP architecture, and that allows to have yet another level of escalation. The MasterManager will finally be the interface with the rest of the system, like the LMC, etc.