DROP: dataIsland, node, location, targetPhase
The 'location' attribute was old, and therefore it has been removed from DROP classes. Instead we have now the 'node' (already present) and 'dataIsland' (new) attributes that point out where the DROP is physically located. These two attributes default to None, as in some cases (like during tests or when feeding physical graphs to DROP managers down in the management hierarchy) this information will not be present in the physical graph.

Regarding the new 'targetPhase' attribute, it indicates the desired level of resilience that the data associated to the DROP should achieved. This attribute should be used by the DLM to create the appropriate replicas of the data.