Fixing Pyro support for sending Events
This was the main cause why the EvtConsumer and friends were still using drop.addConsumer() instead of drop.subscribe(), contradicting the idea of using RCP calls only between DROPs, and using event handlers for the rest. Having an event handler via Pyro meant that an Event would need be sent via Pyro. This wasn't currently supported though, becuase from Pyro >= 4.20 the serpent serializer is used by default, which doesn't allow custom classes to be serialized.

To fix this a custom converter has been introduced that will allow de/serialization of Event objects; code to change to the pickle serializer is also provided but commented out, as such solution is deemed inferior.

With that fix in place, now EvtConsumer has a handleEvent method, and the DROPWaiterCtx and EvtConsumerProxyCtx classes use drop.subscribe().

This exercise also takes us a step closer to the idea of actually sending events between DROPs instead of calling methods on them.