Using events to drive the graph execution.
This is the first take on having events (still sent via RPC calls though) to drive the execution of the graph. This system has very few differences with the direct RPC calls we had before, but proves that an event-based mechanism should work without problems.

Most of the code is unchanged, and even a pair of assertions were added to the unit tests. The only main difference comes in the way the execution mode is handled now that we have events, and how that also affects the integration of the graph with luigi.