Easier Pyro handling of DROP Managers
Before we were creating a Pyro Object ID consisting on the manager acronym (nm, dim or mm) and the DROP Manager ID. This is actually different to handle, making it specially difficult for composite managers to interact with other managers that had already started and hadn't use the same algorithm to create their Pyro Object IDs. We are using now a more standard nomenclature where the Pyro Object ID consists only on the acronym, and thus can always be retrieved by the composite managers, independently of the ID that was given to the sub-manager.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>