Added support for indexed I/O reference
Now inputs and outputs can be referenced by UID in the command-line, both in the dockerApp and in the BashApp. This commit includes the centralization of the command-line parsing code in order to avoid duplicated code between the two apps, plus unit tests to confirm that things are working as expected.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>