Support for node recording in CompositeManager
"Nodes" are different from "dmHosts" in the CompositeManagers, as they point to all the machines that are underneath a given manager, regardless of their degree of separation in the management hierarchy. In the case of the DataIsland manager the list of dmHosts and nodes is the same, but for the MasterManager the list of nodes is potentially bigger depending on how many islands it manages.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>