Support for zeroconf on dfms daemon
The dfms daemon now has optional (default: yes) support for zeroconf-based NodeManager discovery in the network. Daemons running a MM will browse for NM, and each time a NM comes up/down the information is passed down to the MM to keep track of the hosts it knows. This information will be used in the future to be able to accept graphs that don't specify hosts, and thus the MM can assign hosts to it if required.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>