DFMS Daemon simplified and stabilized
The DFMS Daemon now uses the subprocess package instead of the multiprocess
package (i.e., it exec() instead of fork()) which makes things way more simple
since the HTTP server doesn't need to be stopped in the child process anymore.
Now unit tests (of which there are more, covering more functionality) are
really stable, as well as the command-line script.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>