Removing Pyro to expose Drop Managers, take 1
For a while we have been exposing the Drop Managers via two different
interfaces: one based on Pyro, and a different one based on a REST server. We
have successfully used the REST interface to externally interact with the
system, proving that it's reliable and it works, and thus in principle we don't
need the Pyro int erface anymore. Removing it only simplify things.

This commit removes the Pyro exposure of the Drop Managers. Instead they access
each other using the REST client. Minor fixes had to be put in place for this
REST client and server to work as intended, but overall the work was already
done. The MM tests were also found to be actually testing the wrong thing, so
those have been fixed as well.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>