Correct two-way binding of Drops via Pyro
Relationships between Drops are always two-way, and because of that the Drop
methods that establish such relationships have always had an automatic
back-reference mechanism (e.g., when an input is appended to a consumer, the
consumer is appended to the input).

When establishing relationships between Drops that belong to different nodes
the situation is a bit more delicate though: we were getting their Pyro Proxies
from the composite manager (either DIM or MM) and calling the corresponding
method on one of them. What happens then is, for instance:

 * d1 and d2 are proxies, D1 and D2 are the actual objects, and live in N1 and
   N2 respectivelly
 * From composite manager@N0: d1.addInput(d2)
 * In N1: D1.addInput(d2); then d2.addConsumer(D1)
 * In N2: D2.addConsumer(d1)

In principle this scheme works without problems, so far as one doesn't expose
Pyro Daemons to all interfaces (via 0.0.0.0). This is not officially supported
by Pyro actually (see [1]), but we do work around it anyway in order to be able
to interact via more than one interface. They way this is done is simple: in
the composite manager we "sanitize" the URI of each Drop before creating the
Proxies, replacing the "0.0.0.0"s with the correct IP as necessary.

This trick works at the composite manager level alright, but in the example
above d1 is calculated once again to satisfy the back-reference call with D2.
This second time the proxy URI is directly handed over by D1's daemon to the
remote D2 object, and we have no way to intercept that process.

To solve the issue we had to change the way we do the two-way binding. By
default we still continue doing the two-way binding, but at the composite
manager level we explicitly make two independent calls, having full control
over how each object sees each other.

In this commit is also included a small-/mini-/micro-optimization, which is to
avoid the two-way binding mechanism to trigger a third call to the original
method being called, although this had a shortcut anyway to detect this
situations.

[1] https://pythonhosted.org/Pyro4/tipstricks.html#multiple-network-interfaces

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>