New unittest that checks drop creation times
This unit test was created while trying to duplicate a problem we were observing
at AWS when using dfms. Drops were taking simply too long to be created, using a
few milliseconds of CPU, with big ~2 [s] pauses in between. We wanted to make
sure that with big graphs this isn't really the case, and at the same time to
replicate the environment used at AWS by spawning separate processes.

The test successfully runs in my machine, although I set a big threshold on the
expected creation time just to make sure the test passes as well on slower
hardware (but not as big as to hide the issue). In the meanwhile more clues seem
to indicate that the problem was in the AWS infrastructure, but keeping this
test here isn't bad anyway.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>