Event listeners must be context managers
So far we have been using Pyro proxies to perform remote code executions
invisibly from within drops themselves. This is hitting us now in the face,
because we are not properly releasing the proxy connection after performing
some remote invocations, thus preventing the server from serving other
requests.

The only way to fix this is to properly release the connection from the proxy.
This can be done either by explicitly calling proxy._pyroRelease() on the proxy
object, or by using the proxy in a 'with' statement. The first option would
polute the lower layers of the code with pyro-specific "if" statements, while
the second requires that the drop class itself, plus a few other ones,
implement the context management protocol, an otherwise unnecessary
requirement. I finally chose the second option, which seems less harmful and
more generic.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>