Initial version of the user-custom error listener
Users are now able to register custom "error listener"; i.e., objects that are
invoked each time a drop goes into the ERROR state. The object's on_error
method is invoked with the failing drop instance as its unique argument.

The error listener is executed within the same thread that is propagating the
state change of the failing drop; thus it should return quickly.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>