Added graph size query capabilities on DMs
Now DMs can report on the size of the graph held by a given session. This is
supported by the REST API as well, which means that it works recursively in the
whole hierarchy.

Unit tests now also use this feature, and the web UI now shows the number of
drops for each session.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>