Avoiding exhaustion on many relationships
We are running out of Pyro threads when we have many drops using one another
without releasing the corresponding proxies. This means that applications must
be aware of this and make sure they release their input/output drops when they
are not needed anymore. I've done this for the SleepAndCopy app but we need to
go through all applications and make these changes. An application-agnostic
solution would of course be even better, but might require more thinking.

This problem also adds up to the Pyro-related problems we have been
experiencing, bringing up again the discussion on whether we should try
alternative methods of communication, both in terms of other frameworks or
designs.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>