KeyError gone in AWS with inter-node relationships
This error was found by Kevin in AWS when deploying one of his graphs which
contained inter-node drop relationships. The bit that triggered the problem was
that in his PG drops receive different values for OIDs and UIDs, which exposed
a problem we hadn't seen until now: the values stored in the interDMRelations
array (instances of DROPRel) pointed to drops by their OIDs, but later the code
used these values as UIDs, causing a KeyError while accessing a dictionary
indexed by UID.

This patch fixes the problem temporarily, but in a nasty way. A proper solution
is to change the requirements given to the clients. So far we require clients
to give always a OID for a drop in a physical graph, and optionally a UID.
Because of that we don't know really in advance the UID that the drop will have
when actually created. The comment above the fix explains the situation in more
detail.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>