First version of replay manager
The replay manager takes stored data from a previous dfms run (a file with the
physical graph, and a file with snapshots of its status through the time) and
replays it. It exposes the same programatic interface as the other managers
(i.e., extends the DROPManager class) and its REST server also obeys the same
interface as the others (i.e., it extends the ManagerRestServer class). This
meant that no additional changes were required on the client side to use this
new manager.

An additional reset() method is supported by this manager, and exposed through
the REST API, in order to be able to re-visualize a run that has been already
displayed.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>