Not invoking drop methods if not really needed
Docker and bash apps were performing command-line substitution blindly without
considering whether the individual substitutions were really needed or not.
This turned out to be a real problem when the command-line substitution was
triggering a remote invocation in a drop that didn't exist anymore, but that at
the same time wasn't needed because the command-line didn't contain any path or
data URL references.

Now command-line substitution only performs substitution as needed; that is,
only if the required reference is found on the command-line string.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>