Trying to normalize how loggers are used
Instead of doing .isEnabled() and then format the message string with its args
directly using % we should be passing down the message and the arguments to the
logger for it to do all the work. This makes the code easier to read. In some
cases still makes sense to check the logger level first, because what will end
up in the message might cost too much to get (e.g., the logs from the docker
containers).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>