InvalidDropException appearing on the show
This new exception denotates a particular kind of error that only is raised
during drop construction, and thus contains drop-specific information. This
will make it easier to know from the client-side what went wrong when deploying
a session, if anything.

I tried to use it in all places I could find, but there might be a few missing.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>