Parallel relationship establishment revisited
This patch implements what Chen and I discussed in Malta: to group
relationships by drop and dispatch those groups to the thread pool instead of
individual relationships. This should avoid exhausting the pyro daemon threads
on the server (since one drop is accessed only by one thread, and therefore
there are less probability of different threads accessing the same remote
daemon) and reuse the same connection on the client.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>