Using dictionary methods safely
In Python3 the iter* methods of the dict class are gone, and now the normal
methods (e.g., "values") return a view iterator of the values instead of a
copy. This patch addresses both changes, making sure we make a copy of the
returned list if we need to modify it or index over it. If we simply need to
iterate over the elements we use the normal method.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>