Reverting back to Pyro's pickle serializer
The serpent serializer started to fail with python3 (maybe because of the
bytes/bytearray types, but not fully confirmed). Reverting back to pickle fixes
the issue and works both for python 2 and 3, while at the same time brings
better performance. The old serpent setup code is still available in case we
want to use it again.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>