Inform the list of node subscriptions to each NM
The DIM now calculates the list of nodes that each NM has to subscribe to in
order to receive the necessary events to continue with the graph execution.
This information is later on given to each NM so they use it as they see fit
(e.g., subscribing to events pushed by those nodes via zmq).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>