Pyro out, ZMQ in to publish events (1st iter)
By hacking our way around the current state of things, we don't let Pyro
Proxies be event subscribers anymore, while on the other hand we finished
connecting the zmq-based event dispatching for events received at the NM level
(which required us to polish a bit the add_node_subscription interface, will
probably even rename it).

The test_dm.TestDM.test_runGraphOneDOPerDOM test is the only one we have ported
to make sure the functionality is correctly tested. The rest of the system is
probably utterly failing.

Removal of a few trailing whitespaces is also included in this patch.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>